Update modal components
The stock generated modal component is very particular about working
mostly via JS LiveView interop. Sometimes I want to show it without JS
extra stuff and this allows me to use the show attribute for that. If JS
turns out inevitable as this feature is iterated upon we can safely
ignore the change or undo it.