Add event listener to push events from handlers
As suggested in the Elixir Forums we add a window event listener to
trigger events from the handlers. We use it specifically to be able to
close modals from handlers without using the flag to indicate if it's
open like we did before. We'll have to continue researching or check why
this might be a terrible idea.